+++ packages/expo-dev-menu/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- [plugin] Migrate import from @expo/config-plugins to expo/config-plugins and @expo/config-types to expo/config. ([#18855](https://github.com/expo/expo/pull/18855) by [@brentvatne](https://github.com/brentvatne))
+
 ## 1.3.0 — 2022-09-16
 
 ### 💡 Others